sửa giao diện, chức năng xoá 1, xoá check của giahan

+++ app/Http/Controllers/GiaHanController.php
@@ -203,7 +203,7 @@ class GiaHanController extends Controller
       return Redirect::to('/home');
     }
   }
-  public function delete_giahan($ma_gh){
+  public function delete_giahan(Request $request){
     $this->check_login();
     $ma_vc = session()->get('ma_vc');
     $phanquyen_admin = PhanQuyen::where('ma_vc', $ma_vc)
@@ -213,14 +213,37 @@ class GiaHanController extends Controller
       ->where('ma_q', '=', '6')
       ->first();
     if($phanquyen_admin || $phanquyen_qlcttc){
-      $giahan = GiaHan::find($ma_gh);
-      if($giahan->file_gh != ' '){
-        unlink('public/uploads/giahan/'.$giahan->file_gh);
+      if($request->ajax()){
+        $id =$request->id;
+        if($id != null){
+          $giahan = GiaHan::find($id);
+          if($giahan->file_gh != ' '){
+            unlink('public/uploads/giahan/'.$giahan->file_gh);
+          }
+          $giahan->delete();
+        }
       }
-      $giahan->delete();
+    }
+  }
+  public function delete_giahan_check(Request $request){
+    $this->check_login();
+    $ma_vc = session()->get('ma_vc');
+    $phanquyen_admin = PhanQuyen::where('ma_vc', $ma_vc)
+      ->where('ma_q', '=', '5')
+      ->first();
+    $phanquyen_qlcttc = PhanQuyen::where('ma_vc', $ma_vc)
+      ->where('ma_q', '=', '6')
+      ->first();
+    if($phanquyen_admin || $phanquyen_qlcttc){
+      $ma_gh = $request->ma_gh;
+      $list_giahan = GiaHan::whereIn('ma_gh', $ma_gh)->get();
+      foreach ($list_giahan as $key => $giahan) {
+        if($giahan->file_gh != ' '){
+          unlink('public/uploads/giahan/'.$giahan->file_gh);
+        }
+        $giahan->delete();
+      }
       return redirect()->back();
-    }else{
-      return Redirect::to('/home');
     }
   }
   public function delete_all_giahan($ma_l, $ma_vc){
